Refactor out uiNoteHeader and uiNoteComments into separate modules
uiNoteEditor was getting big!
