+++ modules/ui/index.js
@@ -34,7 +34,9 @@ export { uiMapInMap } from './map_in_map';
 export { uiModal } from './modal';
 export { uiModes } from './modes';
 export { uiNotice } from './notice';
+export { uiNoteComments } from './note_comments';
 export { uiNoteEditor } from './note_editor';
+export { uiNoteHeader } from './note_header';
 export { uiPresetEditor } from './preset_editor';
 export { uiPresetIcon } from './preset_icon';
 export { uiPresetList } from './preset_list';
